MAINT: removing test_upwind_coupling.py
